--- a/.gitmodules
+++ b/.gitmodules
@@ -1,3 +1,3 @@
 [submodule "benches/vcl"]
-	path = benches/vcl
+	path = benches/cpp/vcl
 	url = https://github.com/vectorclass/version2