C++ рулит :)

(для плюсов специально выставлены флаги `-Ofast` и `-fveclib=libmvec`, чтобы скалярный код векторизовался)
```
test bench_sin_f32_scalar           ... bench:     948,430 ns/iter (+/- 6,007)
test bench_sin_f32_vec              ... bench:     136,525 ns/iter (+/- 1,474)
...
test vcl_bench::bench_ExpScalar_cpp ... bench:      80,149 ns/iter (+/- 1,237)
test vcl_bench::bench_ExpVCL_cpp    ... bench:      74,194 ns/iter (+/- 1,947)
test vcl_bench::bench_SinScalar_cpp ... bench:      90,799 ns/iter (+/- 1,369)
test vcl_bench::bench_SinVCL_cpp    ... bench:      92,291 ns/iter (+/- 1,074)
```