@NamorNiradnug 2024-05-03
С `-O3` выглядит так:
```
test vcl_bench::bench_ExpScalar_cpp ... bench:     456,992 ns/iter (+/- 9,823)
test vcl_bench::bench_ExpVCL_cpp    ... bench:      80,420 ns/iter (+/- 1,070)
test vcl_bench::bench_SinScalar_cpp ... bench:   1,080,094 ns/iter (+/- 7,431)
test vcl_bench::bench_SinVCL_cpp    ... bench:     106,585 ns/iter (+/- 1,258)
```