@NamorNiradnug 2024-05-04
После отпимизаций `sin/cos` почти сравнялись.
Предполагаю, что оставшийся зазор в ~1% — какие-нибудь лишние проверки, которые без `unsafe`-а не убрать.
```
test bench_cos_f32_core_simd        ... bench:     975,239 ns/iter (+/- 17,079)
test bench_cos_f32_scalar           ... bench:     947,916 ns/iter (+/- 13,282)
test bench_cos_f32_vec              ... bench:     105,525 ns/iter (+/- 1,427)
test bench_cos_f64_core_simd        ... bench:   1,990,004 ns/iter (+/- 59,352)
test bench_cos_f64_scalar           ... bench:   1,958,594 ns/iter (+/- 11,851)
test bench_cos_f64_vec              ... bench:     487,930 ns/iter (+/- 17,176)
test bench_sin_f32_core_simd        ... bench:     976,594 ns/iter (+/- 17,137)
test bench_sin_f32_scalar           ... bench:     950,661 ns/iter (+/- 19,528)
test bench_sin_f32_vec              ... bench:     105,318 ns/iter (+/- 1,614)
test bench_sin_f64_core_simd        ... bench:   1,963,591 ns/iter (+/- 39,791)
test bench_sin_f64_scalar           ... bench:   1,914,041 ns/iter (+/- 20,737)
test bench_sin_f64_vec              ... bench:     481,824 ns/iter (+/- 14,619)
test vcl_bench::bench_ExpScalar_cpp ... bench:     439,979 ns/iter (+/- 6,899)
test vcl_bench::bench_ExpVCL_cpp    ... bench:      77,665 ns/iter (+/- 1,118)
test vcl_bench::bench_SinScalar_cpp ... bench:   1,064,079 ns/iter (+/- 10,545)
test vcl_bench::bench_SinVCL_cpp    ... bench:     103,842 ns/iter (+/- 1,586)
```