@NamorNiradnug 2024-05-04
Посмотрел ассемблерных выхлоп и обраружил, что `round()` генерирует не одну инструкцию `vroundp`, а несколько (какие-то `or` и `and`, видимо чтобы обеспечить `round`-у нужное поведение в граничных случаях). `mul_add` и `trunc` генерируют (на нужном железе) ровно две инструкции: `vfmadd` и `vroundp`.
```
test bench_cos_f32_vec              ... bench:      97,229 ns/iter (+/- 1,511)
...
test bench_cos_f64_vec              ... bench:     478,899 ns/iter (+/- 37,561)
...
test bench_sin_f32_vec              ... bench:      98,584 ns/iter (+/- 1,052)
...
test bench_sin_f64_vec              ... bench:     519,695 ns/iter (+/- 44,201)
...
test vcl_bench::bench_SinVCL_cpp    ... bench:     103,941 ns/iter (+/- 7,792)
```