@NamorNiradnug 2024-05-20
Обгоняет VCL-евский синус на `-Ofast`!
```
test bench_asin_f32_scalar               ... bench:     822,945 ns/iter (+/- 11,606)
test bench_asin_f32_vec                  ... bench:      68,016 ns/iter (+/- 1,463)
test bench_atan_f32_scalar               ... bench:   1,444,918 ns/iter (+/- 11,674)
test bench_atan_f32_vec                  ... bench:      73,772 ns/iter (+/- 4,658)
test bench_cos_f32_core_simd             ... bench:     998,328 ns/iter (+/- 34,261)
test bench_cos_f32_scalar                ... bench:     999,563 ns/iter (+/- 37,323)
test bench_cos_f32_vec                   ... bench:      88,989 ns/iter (+/- 1,004)
test bench_cos_f64_core_simd             ... bench:   2,047,228 ns/iter (+/- 98,166)
test bench_cos_f64_scalar                ... bench:   2,017,689 ns/iter (+/- 59,677)
test bench_cos_f64_vec                   ... bench:     455,939 ns/iter (+/- 86,437)
test bench_sin_f32_core_simd             ... bench:     998,575 ns/iter (+/- 20,798)
test bench_sin_f32_scalar                ... bench:     987,702 ns/iter (+/- 10,921)
test bench_sin_f32_vec                   ... bench:      91,300 ns/iter (+/- 1,303)
test bench_sin_f64_core_simd             ... bench:   2,046,133 ns/iter (+/- 93,884)
test bench_sin_f64_scalar                ... bench:   1,985,718 ns/iter (+/- 29,760)
test bench_sin_f64_vec                   ... bench:     446,279 ns/iter (+/- 14,796)
test vcl_bench::bench_asin_f32_vcl_cpp   ... bench:      69,023 ns/iter (+/- 2,098)
test vcl_bench::bench_atan_f32_vcl_cpp   ... bench:      84,623 ns/iter (+/- 1,691)
test vcl_bench::bench_exp_f32_scalar_cpp ... bench:      81,997 ns/iter (+/- 2,106)
test vcl_bench::bench_exp_f32_vcl_cpp    ... bench:      75,427 ns/iter (+/- 1,688)
test vcl_bench::bench_sin_f32_scalar_cpp ... bench:      90,706 ns/iter (+/- 2,022)
test vcl_bench::bench_sin_f32_vcl_cpp    ... bench:      93,384 ns/iter (+/- 1,396)
```