@NamorNiradnug 2024-05-20
Перехапустил вместе с `exp`:

|   | `portable_simd_addons` | `VCL` | `libmvec`
|---|  :---:  | :---:  | :---: 
|sin|  89,195 | 92,319 | 90,499
|asin| 68,096 | 67,946 | -
|atan| 72,907 | 83,579 | -
|exp|  83,019 | 75,149  | 81,720