@NamorNiradnug 2024-05-21
Оказывается, для разных функций разный оптимальный размер `N` для `Simd<f32, N>` в цисле: для `sin` и `cos` при уменьшении происходит значительная потеря производительности. А для `exp` наоборот.
```
test bench_exp_f32_vec                       ... bench:      66,949 ns/iter (+/- 761)
test cpp_benches::bench_exp_f32_scalar_cpp   ... bench:      76,863 ns/iter (+/- 2,101)
test cpp_benches::bench_exp_f32_vcl          ... bench:      73,446 ns/iter (+/- 1,323)
```